+++ asm/non_matchings/thread3_main/func_8006AC00.s
