+++ asm/non_matchings/thread3_main/func_8006AE2C.s
