+++ asm/non_matchings/thread3_main/func_8006CCF0.s
