+++ src/menu.c
@@ -21,7 +21,7 @@
 #include "textures_sprites.h"
 #include "thread30.h"
 #include "unknown_00BC20.h"
-#include "unknown_06B2B0.h"
+#include "thread3_main.h"
 #include "unknown_078050.h"
 
 /**
