+++ src/video.h
@@ -67,8 +67,8 @@ extern s32 D_800DE7BC;
 
 void init_video(s32 videoModeIndex, OSSched *sc);
 void set_video_mode_index(s32 videoModeIndex);
-s32 get_video_mode_index(void);
-void set_video_width_and_height_from_index(s32 fbIndex);
+UNUSED s32 get_video_mode_index(void);
+UNUSED void set_video_width_and_height_from_index(s32 fbIndex);
 s32 get_video_width_and_height_as_s32(void);
 void init_vi_settings(void);
 void init_framebuffer(s32 index);
@@ -76,9 +76,9 @@ void func_8007A974(void);
 void func_8007AB24(s8 arg0);
 s32 func_8007AB34(void);
 void swap_framebuffers(void);
+void memory_copy(u8 *src, u8 *dest, s32 len);
 
 //Non Matching
-void memory_copy(u8 *src, u8 *dest, s32 len);
 s32 func_8007A98C(s32 arg0);
 
 #endif
